Fix vertical position of owner-drawn wxStaticBox border
There is no need for the extra offset, this makes the border position
different from that drawn by wxStaticBox itself, meaning that the window
layout changes in dark mode compared to the default light one.

This border was added in the initial version of the code in f21bf04a3c
(Draw a grey border around wxStaticBox in dark mode, 2023-02-19) but it
doesn't seem necessary in any of the tested configurations (Windows 7
and 10, with 100%, 150% or 200% DPI scaling).