Simplify SetPropertyColoursToDefault() compatibility handling
The previous version defined different functions depending on whether
3.0 compatibility was turned off or on, which was both unnecessarily
complicated and not really correct, as it prevented the code using the
library compiled with 3.0 compatibility from using the new functions.

Simplify things by always defining the overloads not taking any flags
and taking wxPGPropertyValuesFlags and also defining an additional
overload taking int flags if necessary.

This is much simpler to understand and allows the code using the typed
flags, such as our own propgrid sample, to compile with the library with
3.0 compatibility enabled (except that we don't even need to test for
3.0 compatibility any more, as it's covered by 3.2 test).