Initialize wxPrintDialogData fields in their declaration
This is more clear as it makes the initializer immediately visible and
avoids the need to list all the fields twice, first when declaring and
then when initializing them.

Also use compiler-generated copy ctor and assignment operators to avoid
re-enumerating all the fields again.

No real changes, the same initial values are preserved.