Fix wxKeyEvent::GetKeyCode() for non-US keyboard layouts
Use the key code corresponding to the key in the US keyboard layout for
the key down/up events even when not actually using US layout, as this
is much more useful than simply returning 0 as was done before.

It also is compatible with the behaviour of the other ports.

See #17643, #23379, #23410.

(cherry picked from commit 2c0f6a2aa0c8d0650d1f856aa279ce7495925c17)