Don't assert if HWND is invalid in wxWindow::DissociateHandle()
There was no safe way to prevent wxWindow from trying to destroy the
HWND again if it had been already destroyed, for whatever reason, as
UnsubclassWin() asserted if it was called and calling SetHWND(0) before
calling it as some (but not all) existing code did, resulted in not
removing the association using the old HWND as key.

Allow to use DissociateHandle() for safely detaching HWND from the
window, without assuming that it is valid.