Use DissociateHandle() for the editor destroyed by native control
Don't call SetHWND(0) before calling UnsubclassWin() to prevent it from
asserting if the HWND was invalid because this prevented the association
with the old HWND from being removed from the global gs_windowHandles
and call DissociateHandle() instead.

Don't call SetHWND(0) after calling UnsubclassWin() because this was
useless, as this function already did this anyhow.