Fix using wxComboCtrl popup inside AUI floating panes
Handle correctly the strange case of the popup window being destroyed
by Windows itself, which happened when the AUI floating frame, which was
used as the parent of the popup when it was created for a wxComboCtrl
inside an AUI pane, was itself destroyed. This resulted in problems
because we kept using popup window even though its HWND wasn't valid any
longer.

Detect when this happens and destroy the popup window too, to force
recreating it when/if it's needed again in the future.

This also requires keeping wxOwnerDrawnComboBox::m_initChs even after
using them, as they will be needed again if we need to recreate the
popup.

See #23399.

(cherry picked from commit ac21107b9511fb8239f970d2e5e0a19301176742)