Fix deleting taskbar icon in wxMSW wxNotificationMessage
Previously, the taskbar icon could remain shown in the taskbar after
wxNotificationMessage was destroyed, as wxEVT_TASKBAR_BALLOON_TIMEOUT
handler was called on an already dead object.

Weirdly enough this didn't result in the crashes, but it definitely
didn't work correctly neither.

Fix this by binding a separate handler for this event, not using the
wxNotificationMessage object at all, which is responsible for destroying
the icon when it times out.

This also seems to make it completely unnecessary to have a separate
m_active flag in wxNotificationMessageImpl, as we can now just delete
the "impl" object too directly when the main object itself is deleted.

Closes #23432.