Deprecate useless wxResourceCache class
This is just a trivial wrapper for wxList, which is itself obsolete, so
deprecate it and schedule for removal in wx 3.6.

Also move it entirely inline and don't export it from the DLL any longer
both because this avoids generating any code for it if it's not used (as
should be the case for 99.99% of wx applications) and because we
currently can't use wxDEPRECATED_MSG() and WXDLLIMPEXP_CORE together in
the same declaration.