Use std::vector instead of wxGridStringArray in wxGrid code
This allows to simplify the code by directly using the standard class
methods.

Note that wxGridStringArray is used outside of wxWidgets, so we still
preserve its definition but make it just a simple vector of objects
instead of a legacy object array as there is really no need to keep
pointers to wxArrayString here.