Use wxBaseArray instead of object arrays in wxAuiManager
This is similar to the previous commit and replaces some object arrays
with arrays of objects, as we need to keep providing more or less the
same API for accessing these arrays because they can be used from the
application code, even if they're not formally part of our public API.

Also adjust m_actionPart when removing parts from m_uiParts. This seems
to have been always necessary when deleting the part corresponding to
the action one, but now is also needed when deleting any previous parts
too.