Stop using object array of MSG structs in wxMSW wxEventLoop
This array type doesn't seem to be used anywhere, so remove it entirely,
especially because it allows to avoid including <windows.h> from the
public header.

The array also doesn't even need to be a member variable of
wxGUIEventLoop as it's only used in a single function, so make it a
local variable there.