Stop using object array in wxUniv wxMenuBar implementation
Replace it with a simple vector of wxMenuInfo objects.

Also don't export these objects from the shared library, this is
completely unnecessary.

Note that this required moving wxMenuBar ctor and a method using the
vector directly out of line, as vector of objects of incomplete type
must not be used from the header.