Stop using macro-based wxList in wxArtProvider code
Use std::list<> instead.

Also use std::unique_ptr<> instead of managing memory manually. As this
makes wxArtProvidersList non-copyable, don't export it from the DLL any
longer, as DLL-exported classes apparently must have a copy ctor, or at
least MSVS (all versions up to 2022) try to generate it even if it's
explicitly deleted in the class declaration and fail, due to the base
class being non-copyable. And because there was never any reason to
export neither this class nor wxArtProviderCache from the DLL, stop
exporting the latter one too.

Note that this implies that the only way to delete a wxArtProvider
registered in the global provider list is by removing it from it,
instead of deleting it directly, which was apparently allowed before,
even though it shouldn't have ever been actually done -- but make this
even more clear in the documentation now.