Stop using wxHashMap in wxZipInputStream code
Use std::unordered_map<> instead.

Also use wxFileOffset as the type of the map key instead of long, as
this is not a problem with the standard class and it allows to remove
all the casts from wxFileOffset to long (which were normally lossless
anyhow in 64-bit build, but still ugly).