Use std::unordered_xxx in wxGrid
Make wxGridFixedIndicesSet a std::unordered_set and use unordered_map
for the other maps used by wxGrid.

Remove the initial buckets assignment, as it isn't really clear if it's
helpful.

Also remove the casts that are not needed now that we use a map with the
correct type for the keys and values.