Remove ugly s_fileInfo macro from wxMSW wxFSVolume code
Don't redefine a variable as a macro, this was just horrible.

Also remove a (surely outdated by now, even if it ever was true) comment
about "Cygwin bug": using a wrapper function for a global is a good idea
anyhow as it imposes the correct initialization order, so we don't need
any bogus excuses for doing it.