Use std::unordered_map<std::unique_ptr<>> in wxMemoryFSHandler
Replace a map of raw pointers defined using wx synonym with the direct
use of the standard class using smart pointers to simplify code and make
it more robust.

Don't define wxMemoryFSHash in the global scope any longer, it shouldn't
be used by any existing code, but if it is, it's better to prevent it
from compiling immediately, as it would be broken anyhow by the change
to the map value type.