Use std::unordered_map<> in wxColourDatabase
Stop storing pointers to wxColour in this map when we can just store
wxColour objects themselves.

Also avoid dependency on <unordered_map> in the header by only defining
the map in the source file: as wxColourDatabase already used only a
pointer to the map, rather than the map object, this was easy to do.