Give a debug warning if resource bundle has only one bitmap
Typically, wxBitmapBundle::FromResources() is used from inside
wxBITMAP_BUNDLE_2() macro, which requires having a 2x bitmap version
under other systems, so it should be pretty rare not to have such 2x
version under MSW too and if it does happen, it's quite likely to be the
result of an error.

To make such errors simpler to find, give a debug warning if we don't
find any bitmaps except the main/primary one.