Explain how to specifiy download paths before starting a wxWebRequest
Adding an overview paragraph explaining how to use wxWebRequest's and wxWebEvent's IDs together to specify a request's download path before starting it. The sample shows how to interactively prompt for local paths once a file downloads, but I was wanting to download a file to a specific location without user prompting. The technique described here worked for me.

If accepted, this could close #22986.