Fix updating wxFrame client size in wxEVT_DPI_CHANGED handler
This didn't work correctly when switching from higher to lower DPI if
the frame had a status bar because this status bar still had its old,
bigger height, by the time SetClientSize() was called from the event
handler, resulting in the frame size becoming too big.

It notably resulted in having an unwanted gap between the frame contents
and the status bar if the event handler called GetSizer()->Fit().

Fix this by calling PositionStatusBar(), which also updates the status
bar height to correspond to the new DPI as a side effect, from a new
virtual MSWBeforeDPIChangedEvent() function which had to be added to
allow wxFrame to customize DPI handling.

Adding this new virtual function just for this isn't great, but the only
alternatives seem even worse.