Pass wxDPIChangedEvent to the just added MSWBeforeDPIChangedEvent()
This is unusual, but this function may need to rescale some coordinates
and it's convenient to just call event.Scale() then.

So pass it the event that we are preparing to process to it instead of
passing it new and old DPI separately and forcing it to deal with them
itself (instead of using Scale()).