Fix a memory leak when using non-wx threads
The wxThreadSpecificInfo object created by wxThreadSpecificInfo::Get()
was only released by the wxThreadInternal's cleanup, which meant that it
was leaked if wxThreadSpecificInfo::Get() was called by a thread not
created by wxThread, e.g. when using std::async or std::thread.

As we still can't rely on C++11 thread local variables in this branch
(see d666d1e222 (Fix a memory leak when using non-wx threads,
2023-05-10) in master), we work around the problem by replacing the
ThreadCleanUp call with platform specific implementations for
wxThreadSpecificInfo::Get() that free the allocated wxThreadSpecificInfo
automatically on thread exit.

Note that the Fls functions used for Windows implementation are not
supported on Windows XP, and thus the problem still persists there.

Closes ##23535.