Fix wxMSW wxWindow::GetTextExtent() with given font in high DPI
This function returned wrong results when using a non-default font in
non-standard DPI because it didn't adjust the font to use the window
DPI, as needs to be done manually in wxMSW before using it.

Fix this now and do the same thing here as we already did in
wxWindow::SetFont() (via WXAdjustFontToOwnPPI()) or in wxDC::SetFont().

See #23542.

(cherry picked from commit f3a4767455f62d1a614eac6d538599ac738d3918)