Fix wxStatusBar::GetBackgroundColour() in dark mode under MSW
Get the colour actually used by the theme part we use for the status bar
in the dark mode, which is quite different from the default background
colour (0x333333 vs 0x202020).

This is notably important for the controls that live inside status bar
panes, as they must use the correct background colour to blend in.