Revert to using wxBaseObjectArray for wxAuiPaneInfoArray
This reverts changes to one of the arrays in 39706e4db0 (Use wxBaseArray
instead of object arrays in wxAuiManager, 2023-04-11) as this array is
returned by wxAuiManager::GetPane() and the existing code relies on the
references to the elements of this array remaining stable even if the
array is modified, which would be impossible with an array storing the
objects directly, so return to storing heap-allocated pointers to it.

There should be no such problems with the other arrays changed by that
commit, so leave them unchanged.

Closes #23545.