Improve wxGLCanvasEGL refresh logic under Wayland
Ensure that we only draw when we really should and do not block in
SwapBuffers() when we shouldn't.

Use eglSwapInterval() to disable blocking in eglSwapBuffers() and ensure
that we only call it when we should by using m_readyToDraw not only for
the first repaint, but also for all the subsequent one.

This should result in better performance.