Fix drawing of disabled wxStaticText in dark mode
Disabled static controls looked ugly as greying out the label doesn't
work at all when the background is dark, so override their drawing in
this case and just draw them using the same grey-ish colour that is
already used for the disabled menu items.

This is somewhat of a hack, but doesn't seem to have any bad effects and
is much simpler than using owner-drawn controls.