Inline wxStringFormat Validate in-class
wxStringFormat's Validate method was being provided an empty, inline implementation
where wxDEBUG_LEVEL did not evalue to true, but outside of the class. This resulted
in multiple definitions of the method in different compilation units, which MSVS 2019
did not like.

This change moves the inlined definition to the declaration site with the same
predicate.