Initialize wxDIB members in their declarations and drop Init()
Remove the Init() helper initializing all members as we can now do it in
their declarations instead which is even better.

Also drop Free() which existed mostly for symmetry with Init().

As a side effect, this change allows to include this header from outside
the library (e.g. from the tests code), while previously it would result
in the following warning from gcc:

‘void wxDIB::Init()’ redeclared without dllimport attribute after being
referenced with dll linkage