Fix drawing of the dialog size grip in dark mode
We need to manually adjust the background of this window as it uses
fixed background colour, which is wrong even in the default light mode
if the dialog uses a different background colour, and wrong even with
the default background in dark mode (where the gripper uses #383838
while the default dialog background is #202020).

Reuse the CustomPaint() function which was already used for drawing some
other controls in dark mode to do it.