Add support for getting theme data for non-default DPI too
Use OpenThemeDataForDpi() if necessary to retrieve the correct metrics
values for non-default DPI.

Change the existing wxUxThemeHandle ctor taking wxWindow to use the
window DPI automatically and replace the ctor taking HWND with a pair of
functions either explicitly specifying the DPI or indicating that it is
not used -- the latter can be useful to avoid unnecessary complications
when the theme is only used to retrieve some colour values that don't
depend on the DPI anyhow.

Also make wxUxThemeHandle class movable as a side, but possibly useful,
effect of these changes.