Restore setting CRT locale in wxLocale under macOS
This was accidentally broken by the previous changes and the C-level
locale was not set at all any more, even when using wxLocale (and not
just wxUILocale, which is not supposed to set it) under macOS.

Fix this by explicitly calling setlocale() on this platform to restore
the previous, and documented, behaviour.

See #23557, #23603.

(cherry picked from commit 54a2eb16871e5159f2baa3228d0a6dd70b10633c)