Draw all wxListCtrl items ourselves in dark mode
Unfortunately drawing just the selected item ourselves and let the
control draw all the other ones doesn't work too well as the offsets
used for the items differ between Windows versions and so it's difficult
to get them right, but getting them wrong is very noticeable as when
browsing through the control the current item seems to jump because it's
shifted compared to the other ones.

So just draw all the items ourselves: this is not as nice as letting
Windows do it, but it ensures that they use consistent positions.