FEATURE: Allow the use of custom colours to paint the selection on a wxListCtrl
Normally the system defined colours are used for painting the selected itens on a wxListCtrl.
This change allows using different foreground and background colours for it, conditionally.
The methods implemented for it are:

  void EnableAlternateSelectedColours(bool enable = true);

Choose which colours to use; system-wide (false) or custom (true)

  bool IsEnabledAlternateSelectedColours();

Check if custom colours are being used

  void SetAlternateSelectedTextColour(const wxColour& colour);
  void SetAlternateSelectedBackgroundColour(const wxColour& colour);

Set the foreground and background colours to use instead of the system defined ones

  wxColour GetAlternateSelectedTextColour();
  wxColour GetAlternateSelectedBackgroundColour();

Get the custom foreground and background colours used