FEATURE: Added a new style flag to the wxDataViewCtrl class to prevent interactive searching on wxGTK
As implemented, the wxDataViewCtrl enters in an interactive search mode whenever a keypress is not
handled by the main control. That causes a popup text entry control to be shown below the main control,
to no purpose as the main control does nothing with it.
Adding the new wxDV_PREVENT_SEARCH style flag will prevent that behaviour on wxGTK and will do nothing
on all other ports.